MainVerticleTest, Tenant2ApiTest, ApiIT ordering
After ApiIT was added it was no longer possible to run Tenant2ApiTest
and MainVerticleTest without failures. Solution: reset RestAssured
static initializers. I don't know why ApiIT was an integration test
to be begin with. After this change, it no longer need to be.
So it could be renamed as well.